requirements.txt
streamlit==1.29.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
fastapi
uvicorn[standard]
pydantic
supabase
python-multipart
fastapi-cors